Move Change interfaces to types.ts
